refactor: v1 디렉토리를 생성 및 파일 이동
