+++ product-reservation-service/src/main/java/com/monkey/productreservationservice/application/service/v1/ProductReservationServiceApiV1.java
@@ -1,4 +1,4 @@
-package com.monkey.productreservationservice.application.service;
+package com.monkey.productreservationservice.application.service.v1;
 
 import com.monkey.common_module.aop.AccessLevel;
 import com.monkey.common_module.aop.CheckUserRole;
@@ -9,8 +9,8 @@ import com.monkey.productreservationservice.application.dto.response.ResProductR
 import com.monkey.productreservationservice.application.dto.response.ResProductReservationGetDTOApiV1;
 import com.monkey.productreservationservice.application.dto.response.ResProductReservationPostByIdCancelDTOApiV1;
 import com.monkey.productreservationservice.application.dto.response.ResProductReservationPostDTOApiV1;
-import com.monkey.productreservationservice.application.validator.ProductReservationReadValidator;
-import com.monkey.productreservationservice.application.validator.ProductReservationValidator;
+import com.monkey.productreservationservice.application.validator.v1.ProductReservationReadValidator;
+import com.monkey.productreservationservice.application.validator.v1.ProductReservationValidator;
 import com.monkey.productreservationservice.domain.entity.ProductReservationEntity;
 import com.monkey.productreservationservice.domain.repository.ProductReservationRepository;
 import com.monkey.productreservationservice.domain.vo.ProductReservationStatus;
